Merged PR 9248: Use safe_load for PyYAML
Introduced a model class look-up table that alleviates the need to specify the desired model class in the YAML file. Referencing classes in YAML files requires unsafe loading, which is undesirable from a security stand-point.

Some refactoring along the way, so the code will run without appending "src" to the PYTHONPATH.